backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-generativeai==0.5.4
supabase==2.5.1
python-dotenv==1.0.1
httpx==0.27.0
